Update
Finalized all of the functions to create the word frequency distribution dataframe for each life cycle stage.  Moved all of the functions to a module and then create a pipeline for all of the major funtions.  The next step will be to modify the dataframe to calculate teh average and standard deviation for each stage versus its peers and then identify the top 5 words with the highest standard deviation.