Updates
I created a complete redesigned stage 1 that creates a frequency distribution by stage of Ngrams and is modular so that the user can chose which Ngram they want to use.  The next step is to create an option where the user can chose as to whether or not they want an actual frequency, i.e. the number of times the Ngram appears in a given stage or return the value as a % of times the Ngram appears / total number of rows in the Docketsheet.

Note that the Module1 will need to be cleaned up.

Given the speed increases associated with the new model we can likely combine Stage 1, 2 and 3.