Updates - Stage 1 finalized
Stage 1 has been finalized.  Stage1 code and module have been updated to include a function that transforms the FreqDist dataframe to create a tuple of the Ngrams and remove the individual columns.  In addition, we no longer have our target variable in the index, so the subsecuent steps include: how to address the fact that the index is no longer our target, we have a new column title 0, we have a new object that is a tuple, and we don't understand yet whether or not the thresholds set in the original code actually apply to Ngrams.