Updates Stage3 Code & Module
Stage3 Code & Module are now finished.  The code is modular, can be used to iterate over an entire directory selecting certain methodologies to focus on, write to excel, choose target destination to write files, etc. 

Issue Report:
The KeyPhrase for Nograms is printing an error that it couldn't identify the Ngram type.  Need to fix Module.