Stages 1-3 Updated
There was an error in Stage 2 regarding how column 0 was referenced in the dataframes for the 4 types of Ngrams from Stage1.  Nograms dataframe column 0 needed to be renamed to 0 as opposed to 'index', 

Stage2 also had some minor erorrs in the dataframes that were referenced inside the Code, so that was updated. 

Test stage 1 -3 and all run fine now. 

The only thing I can think of is in Stage 3 we are referencing the file name to obtain the type of Ngrams.  I thought we had to change the file name for Nograms, but I think that is fixed. 

Next Steps:
- Calculate how many columns for each KeyWord do are all zeros.  
- Test Stage 4. 
- Pipe everything together into a single code.  Save everything first before doing so.  Create a stage where they are separate and all the code works.